ytpmv_generator.py
from moviepy.editor import VideoFileClip, AudioFileClip, CompositeVideoClip
from moviepy.video.fx.all import *
from pydub import AudioSegment
import numpy as np
import ffmpeg

class YTPMVGenerator:
    def __init__(self):
        self.video_clip = None
        self.audio_clip = None
    
    def load_source(self, video_path):
        """Load source video file"""
        self.video_clip = VideoFileClip(video_path)
    
    def load_audio(self, audio_path):
        """Load audio track"""
        self.audio_clip = AudioFileClip(audio_path)
    
    def apply_pitch_shift(self, clip, semitones):
        """Simulate Sony Vegas pitch shift effect"""
        return clip.fx(vfx.time_mirror)  # Example effect
    
    def create_stutter_effect(self, clip, interval=0.1):
        """Create stutter effect common in YTPMVs"""
        clips = []
        duration = clip.duration
        current_time = 0
        
        while current_time < duration:
            segment = clip.subclip(current_time, min(current_time + interval, duration))
            clips.append(segment)
            current_time += interval
            
        return concatenate_videoclips(clips)
    
    def apply_color_effect(self, clip):
        """Apply color manipulation effects"""
        return clip.fx(vfx.colorx, 1.5).fx(vfx.mirror_x)
    
    def create_ytpmv(self, output_path, effects_sequence):
        """Generate final YTPMV with specified effects sequence"""
        if not self.video_clip or not self.audio_clip:
            raise ValueError("Video and audio must be loaded first")
        
        # Apply effects based on sequence
        processed_clip = self.video_clip
        for effect in effects_sequence:
            if effect == 'pitch_shift':
                processed_clip = self.apply_pitch_shift(processed_clip, 4)
            elif effect == 'stutter':
                processed_clip = self.create_stutter_effect(processed_clip)
            elif effect == 'color':
                processed_clip = self.apply_color_effect(processed_clip)
        
        # Set audio
        final_clip = processed_clip.set_audio(self.audio_clip)
        
        # Write final video
        final_clip.write_videofile(output_path, 
                                 codec='libx264', 
                                 audio_codec='aac',
                                 fps=30)
