requirements.txt
moviepy>=1.0.3
pydub>=0.25.1
numpy>=1.21.0
pillow>=8.3.1
scipy>=1.7.0
ffmpeg-python>=0.2.0